fix: handle EOFError for cache reading (#15298)
Description of PR
When parallel run is enabled, multiple processes may try to read/write the same cache file, so there will be a small chance that the file is being written by process 1 while process 2 is reading it, which will cause EOFError in process 2. In this case, we will retry reading the file in process 2. If we still get EOFError after some retry attempts, we will return NOTEXIST to overwrite the file.

In the meantime, we should also optimize how we initialize the DUT hosts when parallel run is enabled to reduce the chance of having such cache read issue.

Summary:
Fixes # (issue) Microsoft ADO 30031372

Approach
What is the motivation for this PR?
To prevent EOFError when reading cache file when parallel run is enabled.

How did you do it?
Add retry mechanism and optimize how DUT hosts are initialized when parallel is enabled.

How did you verify/test it?
I ran the updated code and can confirm parallel run is still working as expected.

co-authorized by: jianquanye@microsoft.com