Move the common code from testcases to the fixtures. (#15099)
Description of PR
Summary:
Snappi reboot testcases don't save the config before rebooting the DUT. This causes the DUT to loose all the config, and the test fails with "ARP is not resolved" errors.

This PR addresses this issue by saving the config before any reboot. Also this PR moves the code that is reused in multiple tests to a common code.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>