Capture before warm-reboot device logs (#14980)
What is the motivation for this PR?
Have experienced issues in the reboot tests and were unable to diagnose due to a lack of information in the warm-reboot sequence.

How did you do it?
Copied across the logs that were already preserved on the device across reboots.

How did you verify/test it?
Tested internally on A->B and multl-hop upgrade scenarios.

Signed-off-by: yenlu <yenlu@cisco.com>