[arp_update] Verify that kernel/APPL_DB MAC mismatch is corrected (#14662)
What is the motivation for this PR?
It is possible for kernel neighbor information to fall out of sync with the APPL_DB neighbor table.

How did you do it?
Manually change the APPL_DB neighbor table entry, then verify that the arp_update script is able to flush the out of sync neighbor.

Signed-off-by: yenlu <yenlu@cisco.com>