fix enable-dynamic-buffer issue
Before stopping buffermgrd, need to make sure buffermgrd is running, Otherwise it might cause some side-effect timing issue.  For example: after loading minigraph, buffermgrd will restart, if we run command to stop buffermgrd when buffermgrd is not running, the buffermgrd will not be stopped really, and the consequence is the command starting buffermgrd triggered by loading minigraph in the queue will be executed in the later. Therefore, when the subsequent starting buffermgrd happens between time changing the config db with the dynamic buffer config and setting the buffer model to dynamic, it will raise some error by sai, because dynamic buffer config will be applied to traditional model.