Add dummy ptf.testutils library to skip traffic test in KVM (#15322)
What is the motivation for this PR?
Currently we are using marker "skip_traffic_test" to decide whether to run traffic test or not. Then the marker would pass to a fixture called by each data plane testcase.
This approach is only applied in master branch, which cause a lot of code change in test scripts, and difficult to do cherry pick.

How did you do it?
Based on the marker "skip_traffic_test", use a testcase level pytest hook to get the marker, if it exists, then setattr to a function always return for all testutils function startwith "send" or "verify", and setattr back when test finished.

How did you verify/test it?
Verified in KVM test

Any platform specific information?
Supported testbed topology if it's a new test case?

Signed-off-by: yenlu <yenlu@cisco.com>