Add mgmt branch parameter in pytest-collect-only.yml (#15348)
What is the motivation for this PR?
Currently pytest-collect-only.yml check only support running on self branch, but in some scenarios we need to use it in target branch.

How did you do it?
Add a new parameter mgmt_branch in pytest-collect-only.yml to support running on target branch

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>