Set default kvm_support to False in ptf_runner (#15380)
What is the motivation for this PR?
In PR #12598, we set default kvm_support to True in ptf_runner params, but most of tests do not support traffic test in KVM, to minimize code change in testcases, it's better to set default kvm_support to False

How did you do it?
Set default kvm_support to False and set kvm_support to True in tests currently running ptf_runner

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>