Retry test_psu.py::TestPsuApi::test_power if power not within tolerance (#14788)
* Retry test_power calculated power above 10%

If the test fails on `abs(power - (voltage*current)) < power*0.1` then retry the test. Test is repeated for a maximum of three times until it passes, else it is failed.
This change is introduced to account for the stringent tolerance level of 10% and the errors that might be caused due to time differences reading each of the parameters.

* Retry test_power.py::TestPsuApi::test_power if power above tolerance

If the test fails on `abs(power - (voltage*current)) < power*0.1` then retry the test. The test is repeated for a maximum of three times until it passes.
This change is introduced to account for the stringent tolerance level of 10% and the errors that might be caused due to time differences in reading each of the parameters.

* Add check to detect occurrence of a failure before power calculation

* Resolve pre-commit check issue

* Refactor test_power function to improve readability

* Use wait_until function for retry

Signed-off-by: yenlu <yenlu@cisco.com>