feat: fix pfc_storm flaky (#15418)
Description of PR
Summary:
Fixes # (issue) 301125860

Approach
What is the motivation for this PR?
pfc_gen_t2 is written and using python3 library format. In some switches the default python is not linked to python3 and still refers to python2.7 which leads to unexpected result.

How did you do it?
Explicitly specify to use python3

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>