Merged PR 8150: Add skeleton for creating str4 root fanout.
To bring up the 800G devices, we are adding the 800G root fanout. The new 800G devices can connect to this root fanout and then connect to the servers.

Although adding a new root fanout requires more work than just adding the device and topo, we are still adding this information in first, to ensure the new device information can be placed on the right place.

This change also duplicated a few devices info from str3 to str4. The str3 information is not removed yet, in case blocking some of our current progress before the root fanout is fully ready.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds the initial configuration and setup for creating the str4 root fanout.
- Added `ansible/str4` with host and group definitions for the new str4 environment.
- Created `ansible/group_vars/str4/str4.yml` to define group variables for the str4 environment.
- Added `ansible/files/sonic_str4_devices.csv` to list devices in the str4 environment.
- Created `ansible/files/sonic_str4_pdu_links.csv` to define PDU links for str4 devices.
- Added `ansible/group_vars/str4/secrets.yml` to manage secrets for the str4 environment.