Add a new case to check if config exist (#15080)
Description of PR
Summary:
Fixes # (issue)
There has been an issue that missing exit in the FRR template that caused nht config is missed in the FRR config when vrf is configurated. After fixing the issue in sonic-net/sonic-buildimage#19587, add a new case to verify the 'ip nht resolve-via-default' should be present in FRR config whether vrf is configurated.

Signed-off-by: yenlu <yenlu@cisco.com>