Fix the parameters for AllPortQueueWaterMark Testcase. (#15482)
The testcase test_qos_sai.py::TestQosSai::testQosSaiQWatermarkAllPorts is failing due to much narrower margins in the lossy queue counter case. So this PR attempts to increase the margin, and also handle a case where the required qos.yml key is not present in the TC's params, but in the params for port-speed structure itself.