[dualtor] Fix snmp/* tests failure on fixture teardown
PR#15359 has introduced a "yield" statement inside the for loop of
duthosts which is causing fixture teardown to fail with "Failed: fixture
function has more than one 'yield'" message. Fix is to move "yield"
statement out of this for loop and do config rollback in a seperate for
loop of duthosts.