fix: fix flaky pfc_storm (#15544)
Description of PR
Summary: Fix flaky pfc_storm stom_restored
Fixes # (issue) 30115860

Approach
What is the motivation for this PR?
Currently we detect flaky in detecting storm restore. The reason was because the storm terminated early and restore itself before LogAnalyzer can detect restoration.

As a result, we want to keep this to be stormed long enough.

After the end of each test case, we have stop_storm so it would be fine.

Signed-off-by: Austin Pham <austinpham@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>