fix flaky tests/autorestart/test_container_autorestart.py (#15526)
Description of PR
Summary:
Fixes # (issue) Fixes 30114172

Approach
What is the motivation for this PR?
Increases the threshold timeout for container check for T2 since the BGP neighbor originally was setup to be 360 for T0. However the amount of BGP neighbor is much more comparing to T0.

Upons investigation, this test case were flaky because our bgp were still in connecting status.

Signed-off-by: Austin Pham <austinpham@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>