[dualtor] Fix snmp/* tests failure on fixture teardown (#15529)
Approach
What is the motivation for this PR?
#15359 has introduced a "yield" statement inside the for loop of duthosts which is causing fixture teardown to fail with Failed: fixture function has more than one 'yield' message.

How did you do it?
Move "yield" statement out of this for loop of duthosts and do config rollback in a seperate for loop of duthosts.

How did you verify/test it?
Ran tests under snmp folder and tests are passing on Arista-7260CX3-D108C8 platform.

Any platform specific information?

Signed-off-by: yenlu <yenlu@cisco.com>