Skipping test_static_route on 8122 (#15272)
What is the motivation for this PR?
All 3 tests in test_static_route are failing on 8122.
Tests are failing because “show flowcnt-route stats“ does not show the 1 test pkt that the test has sent.
“show flowcnt-route stats“ does not show the test pkt because counter config itself failed.
Counter config failed because FLOW_COUNTER_CAPABILITY was enabled recently on ASIC/SDK side for 8122 but 'enable_forwarding_route_counter' is not enabled on SONiC/asic_cfg.json on 8122.
'enable_forwarding_route_counter' is not enabled on SONiC/asic_cfg.json on 8122 because of scale limits (cannot scale more than 50k with the current LPM profile).

As the feature is not enabled for this platform, need to skip this testcase

How did you do it?
Added a skip condition for test_static_route for 8122 platform

Type of change 
-Test modification

Back port request 
-202311
-202405

How did you verify/test it?
Ran test_static_router.py on 8122 and verified it was skipped.

Signed-off-by: yenlu <yenlu@cisco.com>