sonic-mgmt: fix t0-isolated-d128u128s2 topo (#15542)
Description of PR
This PR contains the following changes, in order to achieve a functional t0-isolated-d128u128s2 topo:

The field 'bp_interfaces' in the t0-isolated-d128u128s2 yml file is corrected to 'bp_interface' i.e. w/o the 's'.
Added more VMs, to support the larger number of VMs needed for the topo.
Added the missing leaf template file for the topo.
Fixed the synthesis of the MACs used in ansible roles so that it does not error out after 256 interfaces.
Additionally, this PR fixed the 'bp_interfaces' for t0-isolated-d128u128s1 yml file.