Update j2cli to jinjanator. (#15600)
j2cli is not being maintained anymore and will start to fail to work on ubuntu 24.04, because it is trying to load the imp module, which is deprecated now.

However, j2cli is being archived not being maintained anymore. The author recommends using other alternatives that is actively being maintained, such as jinjanator.

Hence, making this change in order to support the latest OS.

Signed-off-by: yenlu <yenlu@cisco.com>