sonic-mgmt: fix port toggle timeout on many ports (#15573)
For topologies leveraging many ports, such as in
the case of t0-isolated-d128u128s2, the timeout
for non-mellanox fixed-chassis devices is a
static value and is too low for the number of
ports being configured.

In contrast, Mellanox devices use a timeout
proportional to the number of ports being
toggled. This change moves fixed-chassis
broadcom devices to use a proportional timeout
as well.

Signed-off-by: yenlu <yenlu@cisco.com>