Fix acl/test_stress_acl.py using bad interface name for ACL table creation
In `acl/test_stress_acl.py`, it attempts to retrieve an interface that
can be used to create a ACL table. DUTs with and without PortChannels
require different methods respectively.

Currently, it checks by filtering with topo. However, some topology
flags can have configurations that have or not have PortChannels, making
topos no longer a sufficient check.

Fix by checking if a PortChannel exists. If it does - use it. If it does
not - fallback on the secondary method to retrieve a normal interface
name if its a not a dualtor topo (due to
https://github.com/sonic-net/sonic-mgmt/pull/6960).