Merged PR 8324: Ignore ACMS bootstrap error in loganalyzer
# Why I did it
We updated the ACMS bootstrap process, which now times out after one hour. The Loganalyzer detected this error, causing the nightly test to fail.

# How I did it
This error is expected, so we should ignore this error in loganalyzer.

# How to verify it
Run end-to-end test

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request updates the log analyzer to ignore ACMS bootstrap errors.
- Modified `ansible/roles/test/files/tools/loganalyzer/loganalyzer_common_ignore.txt` to add a rule for ignoring ACMS bootstrap errors.