Enforce cross-feature dependency checker in pipeline (#15692)
What is the motivation for this PR?
In PR #15559, we introduced a checker to identify cross-feature dependencies within our repository. At the time, since some dependencies still existed, the checker was configured to only run the script without enforcing any pipeline failures. Now that all cross-feature dependencies have been eliminated, we’ve updated the checker to capture the script's return value and trigger a pipeline failure if any cross-feature dependencies are detected.

How did you do it?
We’ve updated the checker to capture the script's return value and trigger a pipeline failure if any cross-feature dependencies are detected.

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>