Increase interface brinup-up wait time on s6000 testbeds (#15610)
What is the motivation for this PR?
Primary motivation of this PR is to validate the interfaces after the reboot.

How did you do it?
The timeout value for FORCE10-S6000 HWSKU was increased.

How did you verify/test it?
Verified running tests on FORCE10-S6000 testbed.

Any platform specific information?
This change specifically affects the FORCE10-S6100 device configuration.

Signed-off-by: yenlu <yenlu@cisco.com>