[CHASSIS][Voq][QoS]Increasing LACP timer for lag ports for broadcom-dnx neighbor EOS host    (#14469)
escription of PR
Intermittently testQosSaiLossyQueue tests fails due to Port-channel flap on broadcom-dnx T2 Voq chassis.
The reason the port-channel goes down is because this test requires disabling TX on the egress port (which is a member of a port-channel)
With the huge buffer-size, it takes a longer time to send packets . This will result in the TX LACP packets to stop egressing, so after 3 LACP packets are missed (~90s) on the server side the LAG is torn down.

Issue # #11682
Summary:
Fixes # (issue)

What is the motivation for this PR?
Intermittently testQosSaiLossyQueue tests fails due to Port-channel flap

How did you do it?
The lacp timer multiplier on the EOS host is configurable.
By default, timeout is 30 secs with a failure tolerance of 3.
We changed the multiplier to an increased value to hold the connectivity for some time until all packets are sent.
And revert the changes after test case execution.

How did you verify/test it?
Executed qos test cases and verfiy the results.

Signed-off-by: yenlu <yenlu@cisco.com>