Ignore SAI switch register read and write not handled logs (#15737)
Description of PR
This PR relaxes the loganalyzer ignore rules to ignore all "SAI_SWITCH_ATTR_REGISTER_WRITE is not handled" and "SAI_SWITCH_ATTR_REGISTER_READ is not handled" errors.

Summary:
Fixes #15736

What is the motivation for this PR?
To help ignore loganalyzer error messages that lead to testcases failing in sonic-mgmt.

How did you do it?
Modify the loganalyzer regex to not check for GBSAI

How did you verify/test it?
Ran the test to verify it ignores these errors.

Signed-off-by: yenlu <yenlu@cisco.com>