T2-snappi: Split udp stream to 6 ports for lossy, but only one for lossless. (#15698)
Description of PR
Summary:
The test: test_pfc_pause_single_lossy_prio is resulting in flaky results. On inspecting the failed state, we find that the cisco-8000 backplane load-balancing is the cause for flakiness. It needs six streams to have higher chance of traffic being equally distributed in the backplane, and the traffic can be sent without being dropped.

Approach
What is the motivation for this PR?
Flakiness of test_pfc_pause_single_lossy_prio test.

How did you do it?
Changed to 6 udp streams for lossy traffic.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>