Add container length check (#15758)
What is the motivation for this PR?
To enhance error messaging for this particular edge case, as it was previously difficult to discern why the container would be created successfully, but then consistently fail to start.

How did you do it?
Adds a check to prevent containers from being created with name lengths greater than 64 characters, as these containers will not be able to start.

How did you verify/test it?
Ran the setup script, with both a valid name and an invalid name (valid is on the left, invalid is on the right):

Signed-off-by: yenlu <yenlu@cisco.com>