fix: add concurrent run for config_reload pc_test_lag (#15721)
Description of PR
Currently, pc/test_lag_2 is having timeout in teardown with the default timeout of 300 seconds.

Since by default the thread_count = 2, which means at maximum, only 2 available thread can be used to run config_reload in DUT. Which means only 2 DUT are config_reload at a time.

This will be slow and exceed 300 seconds.

Approach
What is the motivation for this PR?
How did you do it?
we should specify the thread_count to be the same number of duthost so that they're all config_reload in concurrent.

Signed-off-by: Austin Pham <austinpham@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>