[Bugfix] Fix missing imports for platform_api_conn in tests/platform_tests/api (#15807)
What is the motivation for this PR?
In PR #15605, the platform_api_conn fixture was relocated to a common location. However, some imports were overlooked in the scripts within the tests/platform_tests/api folder, leading to test failures. This PR addresses the issue by adding the missing imports to resolve the errors.

How did you do it?
This PR addresses the issue by adding the missing imports to resolve the errors.

Signed-off-by: yenlu <yenlu@cisco.com>