fix: psu key error (#15734)
Description of PR
We're getting KeyError: 'led_status', error:'led_status' for everytime the wait_until function runs. This is expected since the information was not populated on the database at the time of checking. When testing manually, it's expected around 1 minute for T2 Supervisor to have data in its database.

Summary:
Fixes # (issue) 30114182

Approach
What is the motivation for this PR?
How did you do it?
This PR changes the behaviour that we update the logic to check the key in the dictionary before getting it to avoid KeyError. And also we update the wait to 90 seconds.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>