Add a delay for counter update interval in flow_counter_utils.py (#15499)
We have observed counter check failure in the static route test within the RouteFlowCounterTestContext.
The root cause was the counters were not updated when they were checked due the counter update interval was not completely elapsed.
So we need to wait for the interval before checking the counters.

- How did you verify/test it?
Run the static route test in a loop, no more counter failures.

Signed-off-by: yenlu <yenlu@cisco.com>