Skipped test_gnmi_configdb.py for multi-asic and T2 (#15883)
Description of PR
<Skipped gnmi/test_gnmi_configdb.py for T2 and multi-asic platforms>

Approach
What is the motivation for this PR?
The test currently does not support multi-asic platforms and was failing for 202405 image on T2 chassis, so currently skipping this test.
ADO #30341886

How did you do it?
By adding a condition in tests_mark_conditions.yaml

How did you verify/test it?
Ran it on T2 Arista Chassis

Co-authored-by: yatishkoul <yatishkoul@gmail.com>
Signed-off-by: yenlu <yenlu@cisco.com>