Refactor function get_iface_ip to common location (#15893)
What is the motivation for this PR?
In PR #15619, a cross-feature dependency was introduced in tests/acl/test_acl.py due to the usage of the get_iface_ip function. To address this issue and streamline the code structure, we have refactored the function and moved it to a common location.

How did you do it?
To address this issue and streamline the code structure, we have refactored the function and moved it to a common location.

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>