Merged PR 8598: Update upgrade pipeline redeploy condition
Currently, when we choose Skip Locking Testbed, this step is skipped. However, this can cause subsequent tests to fail, specifically the step "remove topo for ceos, add topo for vsonic". A new condition has been added to ensure that the pipeline will always redeploy, even if Skip Locking Testbed is selected. This guarantees that the environment will be correctly prepared for subsequent steps.

----
#### AI description  (iteration 1)
#### PR Classification
Code update to modify redeploy condition in the upgrade pipeline.

#### PR Summary
This pull request updates the redeploy condition in the upgrade pipeline to include an additional parameter check.
- Changes in `/.azure-pipelines/sonic-metadata/metadata_upgrade_template.yml` to modify the condition for redeploying the testbed.