Add topology mark into test_gnoi_killprocess.py (#15925)
Description of PR
In PR #12478, a new test script was introduced without the required pytest.mark.topology marker to specify the topology it supports. This PR resolves the issue by adding the missing marker to the script.

Approach
What is the motivation for this PR?
In PR #12478, a new test script was introduced without the required pytest.mark.topology marker to specify the topology it supports. This PR resolves the issue by adding the missing marker to the script.

How did you do it?
This PR resolves the issue by adding the missing marker to the script.

Signed-off-by: yenlu <yenlu@cisco.com>