Temporarily skip qos and pfcwd test scripts in PR testing (#15897)
What is the motivation for this PR?
Since the test scripts for qos and pfcwd are not yet included in PR testing, we have temporarily applied conditional marks to skip them. Once these scripts are integrated into PR testing, this PR can be reverted.

How did you do it?

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>