Some of the Line cards does not support speed change during runtime - check in teardown (#15963)
Description of PR
Some of the Line cards does not support speed change during runtime. PR #15761 took care of it in the testcase but there is a blanket speed change in the teardown which causes test errors. This PR checks that capability both in testcase and teardown fixture.

Summary:
Fixes # (issue)

Approach
What is the motivation for this PR?
Speed change in teardown is causing testcase failures for LCs that do not support runtime speed change

How did you do it?
Check and skip for unsupported LCs

How did you verify/test it?
Verified it on T2 setup

Any platform specific information?
Yes

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>