If we can't get the warmboot-finalizer state, don't exit
In some cases, connecting to the device may fail for whatever reason. If
this happens, the current code returns an empty string as the
warmboot-finalizer state. There's a for-loop that exits when the
warmboot-finalizer state is no longer active. An empty string being
returned here breaks that.

To avoid that, don't exit this loop if we get an empty string.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>