chore: bump BFD toggle wait timeout (#15939)
Description of PR
Bump the BFD toggle wait timeout to 450 seconds to give BFD session enough time to go up/down.

Summary:
Fixes # (issue) Microsoft ADO 30112171

Approach
What is the motivation for this PR?
We noticed that 300 seconds is not enough for BFD session to go up/down, so Cisco suggested us to bump it to 450 seconds.

Signed-off-by: yenlu <yenlu@cisco.com>