Minor changes for snappi_tests for T2-ixia, change udp port calculations. (#15906)
Description of PR
Currently the traffic for RDMA tests use randomly selected udp ports. This sometimes ends up sending traffic in a non-load-balanced way, and causes the tests to fail.

In this PR:

For multiple scripts that use udp: We serialize the UDP port numbers instead of randomly picking the numbers. This allows for more uniform load-balancing spread, and allows tests to pass more.
in multi_lossless helper: We reduce the background traffic rate, and raise the rate for lossless traffic, thus keeping the overall rate still same. This allows for accounting the speed of backplane.

co-authorized by: jianquanye@microsoft.com