chore: toggle plt_reboot_ctrl_overwrite to increase timeout for t2 (#15951)
Description of PR
Summary:
Fixes # (issue) 3045798

The decision was to enable this test gap in #13476 however we mentioned that we need to increase the time.

By default, the reboot time will be decided from constant variable reboot_ctrl_dict which will set the REBOOT_TYPE_COLD = 300

https://github.com/sonic-net/sonic-mgmt/blob/master/tests/common/reboot.py#L133

For T2, this time is not sufficient enough and lead to failure. We will adjust the time accordingly.

This PR will fix that

Approach
What is the motivation for this PR?
How did you do it?
Enable plt_reboot_ctrl_overwrite if platform is T2

Signed-off-by: Austin Pham <austinpham@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>