Fix laganalyzer error in test_iface_namingmode
Similar issue to https://github.com/sonic-net/sonic-mgmt/pull/12174 in that counters
stats continue to be polled in an unstable state.  In the test_iface_namingmode.py
case, TestConfigInterface resets the interface configurations at the end, and these
syslog errors only appear after the test begins the teardown process.

Adding a loganalyzer ignore for this message to this test to fix.