Add debug log in decap test to address no vxlan sender IP issue (#16038)
What is the motivation for this PR?
In PR test, decap test is flaky and have a chance to fail in restart ferret in ptf for ferret: ERROR (not running)
Digging into the issue, I found ptf restart ferret failed because VxLan Sender IP is empty, but we don't have enough debug message to address the issue.

How did you do it?
Add debug log in decap test to findout why VxLan Sender IP is empty.

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>