Add stable test scripts into PR checkers. (#16040)
What is the motivation for this PR?
In PR #15907, we initially added several test scripts to the onboarding PR checkers. After monitoring their performance over a few days, some of these scripts have proven to be stable. In this PR, we promote these stable scripts from the onboarding PR checkers to the standard PR checkers.

How did you do it?
In this PR, we promote these stable scripts from the onboarding PR checkers to the standard PR checkers.

Signed-off-by: yenlu <yenlu@cisco.com>