Add ERR log ignore for ECMP/LAG hash not supported
On Broadcom platforms, ECMP/LAG hash is not supported. Adding ERR log
ignore as it is expected for this error log to print.

Add by extending log regex in individual test cases. Not adding to
loganalyzer_common_ignore.txt because that would cause expected fail
tests to pass during CI.