Add fixture to disable IPv6 for PTF tests (#16153)
What is the motivation for this PR?
To address the unexpected packet drops on the DUT when running the testQosSaiHeadroomPoolSize, we identified that the unexpected packets are sent from the PTF container, and the leaf fanout ARP update is forwarding them to the DUT. Therefore, a fixture has been added to disable IPv6 on the PTF container to prevent the unexpected packets, and the values are restored after the tests.

How did you do it?
Add fixture to disable IPv6 for all interfaces in ptf container before the test and restore the original IPv6 settings after the test.

How did you verify/test it?
Verify it by running tests/logs/qos/test_qos_sai.py::TestQosSai::testQosSaiHeadroomPoolSize[single_asic]

-------------------------------------- generated xml file: /data/sonic-mgmt-int/tests/logs/qos/test_qos_sai.py::TestQosSai::testQosSaiHeadroomPoolSize[single_asic].xml -------------------
------------------------------------------------------------------------------------------- live log sessionfinish ------------------------------------------------------------------------
10:18:37 __init__.pytest_terminal_summary         L0067 INFO   | Can not get Allure report URL. Please check logs
================================================================================= 1 passed, 1 warning in 2604.32s (0:43:24) ===============================================================
Any platform specific information?
str3-7060x6-64pe-1

Supported testbed topology if it's a new test case?
t0-standalone-32

Signed-off-by: yenlu <yenlu@cisco.com>