Merged PR 8848: [Failure Analyzer] Add UT Framework
## Overview
PR to add a unit testing framework for the test failure analyzer script.

`conftest.py` handles cli argument parsing, and `test_sanity.py` provides an example regarding the use of these arguments, along with any relevant logs (along with ensuring that there are no issues with the arguments ingested).

This PR also adds a pipeline, which is run when failure analyser code is modified, to run the unit tests found in the failure_analyzer directory.

## Test Evidence

- Full pipeline run: https://dev.azure.com/mssonic/internal/_build/results?buildId=726402&view=results

- Pipeline run when failure analyzer code changes are not detected: https://dev.azure.com/mssonic/internal/_build/results?buildId=726359&view=results
----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces a unit test framework for the test failure analyzer.
- Added `test_failure_analyzer_unit_tests.yml` to define CI pipeline for running unit tests.
- Updated `requirements.txt` with necessary dependencies for testing.
- Added `conftest.py` to configure pytest fixtures and options.
- Added `test_sanity.py` to include basic sanity tests for pytest arguments.
- Updated `README.md` to reflect changes in the test failure analyzer description.