Add solution for changing port_config.ini at fanout switch according to assigned hwsku (#14388)
- What is the motivation for this PR?
During deploying the SONiC fanout switch, when the fanout switch type is different with the dut switch type. The hwsku and port_config.ini supported in fanout switch usually could not match the requirement of dut. Then a fake hwsku and port_config.ini file is needed to meet this requirement.

- How did you do it?
Add two way to meet the requirement:
Generate a fake hwsku and its port_config.ini
Assign a port_config.ini file

- How did you verify/test it?
Run it in internal regression

Signed-off-by: yenlu <yenlu@cisco.com>