Skipping enable_queue_counterpoll_type fixture for Marvell platforms (#16019)
What is the motivation for this PR?
Fixture enable_queue_counterpoll_type in snmp is enabling queue counterpolls. But for Nokia 7215 armhf platforms, as queue statistics are not supported, there is an error while executing this command : "SAI_OBJECT_TYPE_QUEUE is not Supported". So, added a skip condition inside the fixture for Nokia 7215 armhf platforms

How did you do it?
Added condition inside fixture to skip for Nokia 7215 armhf platform

How did you verify/test it?
Run snmp test suite and verify fixture is not executed on Nokia-7215 armhf platforms