Add remain control plane test scripts into onboarding PR checker (#16229)
What is the motivation for this PR?
Some newly added control plane test scripts were not included in the PR checker. In this PR, we incorporate them into the onboarding PR checkers.

How did you do it?
In this PR, we incorporate the newly added control plane test scripts into the onboarding PR checkers.

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>