Ignore gnmi client stderr (#16259)
What is the motivation for this PR?
We have detected a syntax warning when using the gNMI client with Python 3:

E Exception: error:/root/gnxi/gnmi_cli_py/py_gnmicli.py:508: SyntaxWarning: "is not" with a literal. Did you mean "!="?
E if filter_event_regex is not "":
E /root/gnxi/gnmi_cli_py/py_gnmicli.py:510: SyntaxWarning: "is not" with a literal. Did you mean "!="?
E if len(match) is not 0:

How did you do it?
Do not check stderr in the test

How did you verify/test it?
Run gnmi end to end test

Signed-off-by: yenlu <yenlu@cisco.com>