Updating techsupport commands to include optimized bgp neighbor commands (#15466)
What is the motivation for this PR?
To support covering show techsupport after the enhancement introduced in sonic-net/sonic-utilities#3605

How did you do it?
Used regex to modify the validations to align with the new optimized vtysh command in the show techsupport

How did you verify/test it?
Running the test with sonic-utilities changes

Signed-off-by: yenlu <yenlu@cisco.com>