fix: bump memory limit for Cisco 8000 supervisor (#16206)
During recent nightly runs, we observed that the Cisco 8000 supervisor had an average memory usage of 59.7% (calculated from the values 60.3, 59.9, 58.9, 59.2, 59.8, and 60.2). Since the memory threshold is set at 60%, this resulted in two failures. To ensure the stability of the tests, we propose increasing the memory threshold for the Cisco 8000 supervisor to 65%.

Signed-off-by: yenlu <yenlu@cisco.com>