route-bgp-flap: Fix (#15962)
What is the motivation for this PR?
'test_default_route_with_bgp_flap' fails with "Default route nexthops doesn't match the testbed topology" issue.
The count of the next hops is not matching with the upstream neighbor as per the topology.
How did you do it?
Instead of skipping T3 neighbors of AZNGHub, use the address family and next hops information to find out the list of upstream neighbors.
How did you verify/test it?
Ran the above-mentioned test case on a T2 chassis and made sure the test passed without any issues.

Signed-off-by: yenlu <yenlu@cisco.com>