fix: bump memory limit for Cisco 8000 supervisor (#16206) (#16320)
Description of PR
Increase the memory threshold for the Cisco 8000 supervisor to 65% to make the test stable.

Summary:
Fixes # (issue) Microsoft ADO 30114189

Approach
What is the motivation for this PR?
During recent nightly runs, we observed that the Cisco 8000 supervisor had an average memory usage of 59.7% (calculated from the values 60.3, 59.9, 58.9, 59.2, 59.8, and 60.2). Since the memory threshold is set at 60%, this resulted in two failures. To ensure the stability of the tests, we propose increasing the memory threshold for the Cisco 8000 supervisor to 65%.

How did you do it?
How did you verify/test it?
I ran the updated code on Cisco 8000 and can confirm it's working well.

Any platform specific information?
Cisco 8000 chassis

co-authorized by: jianquanye@microsoft.com