Add additional logging to script `calculate_instance_number.py` (#16331)
What is the motivation for this PR?
In this PR, we have added additional logging to the calculate_instance_number.py script to improve tracking and visibility of its operations.

How did you do it?
Add additional logging to the calculate_instance_number.py script.

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>