Add a timeout for acquiring dpkg lock. (#16328)
What is the motivation for this PR?
In our Impacted Area Based PR testing, we noticed some failure beacuse of the error Could not get lock /var/lib/dpkg/lock-frontend. It is held by process 5166 (dpkg). In this PR, we add a timeout for acquiring dpkg lock to avoid this issue.

How did you do it?
In this PR, we add a timeout for acquiring dpkg lock to avoid this issue.

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>