Add pfc-timer-set to 500mS for pfcwd tests. (#16159)
Description of PR
Summary:
Fixes the flakiness of pfc_gen in pfcwd scripts for cisco-8000. We use a new debug CLI script to force the DUT to wait longer in case of a miss in pfc packets from the fanout due to pfc_gen script. So even if the pfc_gen/fanout misses a couple of pfc frames to DUT, the dut would still not send out data packets.

Approach
What is the motivation for this PR?
Flakiness of pfc-gen. Particularly with 400G links.

How did you do it?
We have added a new dshell based script that will force the DUT to wait before transmitting data in case of a miss in pfc pause frames.

How did you verify/test it?
Ran on our duts, with 100G and 400G.

Any platform specific information?
The new fix specific only to cisco-8000.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>