[Impacted Area Based PR testing] Fix script timing with recent test plans (#16437)
What is the motivation for this PR?
In the previous implementation, we retrieved 5 test case entries from Kusto to calculate the total running time of a script. However, the query included the condition | where FilePath == '{script}', and since each script may contain multiple test cases, this approach only sampled 5 random test cases from the script. As a result, the calculated average running time was inaccurate. In this PR, we resolve the issue by selecting the most recent 5 test plans instead, ensuring that we account for all test cases executed in the script.

How did you do it?
In this PR, we resolve the issue by selecting the most recent 5 test plans instead, ensuring that we account for all test cases executed in the script.