Skip dualtor_io/test_tor_bgp_failure.py (#16449)
What is the motivation for this PR?
The script dualtor_io/test_tor_bgp_failure.py has recently been failing frequently on the KVM testbed. This issue is being tracked in #16448. As a temporary measure, we have applied a conditional mark to skip this test in PR checks.

How did you do it?
As a temporary measure, we have applied a conditional mark to skip this test in PR checks.