Merged PR 9153: Add a pipeline for synching Github branches to sonic-mgmt-int
Add a pipeline for synching Github branches to sonic-mgmt-int

For the AI and Chassis project, we have separate Github sonic-mgmt repo Azure/sonic-mgmt.msft.

There are some branches on this repo also need to be synched to the sonic-mgmt-int repo.

The original pipeline "merge_sonic_mgmt.yml" assumes that the source repo is always  sonic-net/sonic-mgmt. This assumption makes it not work for the new model.

This change introduced a new yaml for such purpose. This yaml supports specifying source repo URL. The target branch is still sonic-mgmt-int.

This new yaml uses the matrix strategy to support multiple synching jobs. The synching jobs can be specified in the parameter `SynchJobs` object.

Initially only synching job for the 202405 chassis branch is specified. If this pipeline works well, the jobs covered by `merge_sonic_mgmt.yml` will be migrated to this pipeline too.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds a new pipeline for synchronizing GitHub branches to the `sonic-mgmt-int` repository.
- Added `/.azure-pipelines/repo_mgmt/branch-sync.yml` to define the pipeline for syncing and merging branches on a scheduled basis.